+++ src/com/mideas/rpg/v2/game/item/stuff/Stuff.java
@@ -109,7 +109,7 @@ public class Stuff extends Item {
 	private final static String strengthStrings = " Strength";
 	
 	public Stuff(int id) { //stuff created when waiting for the server to respond
-		super(id, empty, ItemType.STUFF, empty, 0, 0, 1, 1);
+		super(id, empty, ItemType.STUFF, empty, (byte)0, 0, 1, 1);
 		this.sprite_id = empty;
 	}
 
@@ -142,7 +142,7 @@ public class Stuff extends Item {
 		this.isLoaded = true;
 	}
 	
-	public Stuff(StuffType type, ClassType[] classType, String sprite_id, int id, String name, int quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, Wear wear, int critical, int strength, int stamina, int armor, int mana, int sellPrice) {
+	public Stuff(StuffType type, ClassType[] classType, String sprite_id, int id, String name, byte quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, Wear wear, int critical, int strength, int stamina, int armor, int mana, int sellPrice) {
 		super(id, sprite_id, ItemType.STUFF, name, quality, sellPrice, 1, 1);
 		this.gemBonusValue = gemBonusValue;
 		this.gemBonusType = gemBonusType;
@@ -193,7 +193,7 @@ public class Stuff extends Item {
 		this.isLoaded = true;
 	}
 	
-	public Stuff(int id, String name, String sprite_id, ClassType[] classType, WeaponType weaponType, WeaponSlot weaponSlot, int quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, int armor, int stamina, int mana, int critical, int strength, int sellPrice) {
+	public Stuff(int id, String name, String sprite_id, ClassType[] classType, WeaponType weaponType, WeaponSlot weaponSlot, byte quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, int armor, int stamina, int mana, int critical, int strength, int sellPrice) {
 		super(id, sprite_id, ItemType.WEAPON, name, quality, sellPrice, 1, 1);
 		this.gemBonusValue = gemBonusValue;
 		this.gemBonusType = gemBonusType;
